back to old files

+++ libraries/classes/Plugins/Schema/TableStats.php
@@ -1,6 +1,9 @@
 <?php
+/* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
  * Contains abstract class to hold table preferences/statistics
+ *
+ * @package PhpMyAdmin
  */
 declare(strict_types=1);
 
@@ -11,11 +14,7 @@ use PhpMyAdmin\Font;
 use PhpMyAdmin\Index;
 use PhpMyAdmin\Relation;
 use PhpMyAdmin\Util;
-use function array_flip;
-use function array_keys;
-use function array_merge;
 use function rawurldecode;
-use function sprintf;
 
 /**
  * Table preferences/statistics
@@ -23,6 +22,7 @@ use function sprintf;
  * This class preserves the table co-ordinates,fields
  * and helps in drawing/generating the tables.
  *
+ * @package PhpMyAdmin
  * @abstract
  */
 abstract class TableStats
@@ -42,21 +42,27 @@ abstract class TableStats
     public $heightCell = 0;
     protected $offline;
 
-    /** @var Relation */
+    /**
+     * @var Relation
+     */
     protected $relation;
 
-    /** @var Font */
+    /**
+     * @var Font
+     */
     protected $font;
 
     /**
+     * Constructor
+     *
      * @param Pdf\Pdf|Svg\Svg|Eps\Eps|Dia\Dia|Pdf\Pdf $diagram        schema diagram
      * @param string                                  $db             current db name
-     * @param int                                     $pageNumber     current page number (from the
+     * @param integer                                 $pageNumber     current page number (from the
      *                                                                $cfg['Servers'][$i]['table_coords'] table)
      * @param string                                  $tableName      table name
-     * @param bool                                    $showKeys       whether to display keys or not
-     * @param bool                                    $tableDimension whether to display table position or not
-     * @param bool                                    $offline        whether the coordinates are sent
+     * @param boolean                                 $showKeys       whether to display keys or not
+     * @param boolean                                 $tableDimension whether to display table position or not
+     * @param boolean                                 $offline        whether the coordinates are sent
      *                                                                from the browser
      */
     public function __construct(
@@ -130,7 +136,6 @@ abstract class TableStats
      * Displays an error when the table cannot be found.
      *
      * @return void
-     *
      * @abstract
      */
     abstract protected function showMissingTableError();
@@ -147,8 +152,8 @@ abstract class TableStats
                 $db = rawurldecode($_POST['t_db'][$key]);
                 $tbl = rawurldecode($_POST['t_tbl'][$key]);
                 if ($this->db . '.' . $this->tableName === $db . '.' . $tbl) {
-                    $this->x = (float) $_POST['t_x'][$key];
-                    $this->y = (float) $_POST['t_y'][$key];
+                    $this->x = (double) $_POST['t_x'][$key];
+                    $this->y = (double) $_POST['t_y'][$key];
                     break;
                 }
             }
